src/generator2/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
exceptiongroup==1.2.2
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.4
idna==3.10
jsonschema==4.23.0
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.10.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.6.0
openapi3-parser==1.1.19
packaging==24.2
pathable==0.4.3
pathspec==0.12.1
platformdirs==4.3.6
prance==23.6.21.0
pre-commit==3.8.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.7.1
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.1
